requirements.txt
pytest
requests
requests-mock
python-dotenv # Added to load .env file
pytest-benchmark # For performance tests
stripe # Official Stripe Python library